+++ Quaver/App/Controller/home.php
@@ -7,5 +7,6 @@
 
 namespace Quaver\App\Controller;
 
+$this->addTwigVars('siteTitle', "Welcome to Quaver" . ' - ' . BRAND_NAME);
 $template = $this->twig->loadTemplate('home.twig');
 echo $template->render($this->twigVars);