+++ Quaver/App/Controller/logout.php
@@ -11,9 +11,9 @@ if ($_user->logged){
 	$_user->unsetCookie();	
 }
 
-if (!empty($_GET['ref'])) {
+if (isset($_GET['ref'])) {
     $goTo = strip_tags(addslashes($_GET['ref']));
-} elseif (!empty($_SERVER['HTTP_REFERER'])) {
+} elseif (isset($_SERVER['HTTP_REFERER'])) {
     $goTo = strip_tags(addslashes($_SERVER['HTTP_REFERER']));
 } else {
     $goTo = '/';