+++ Quaver/Config.php
@@ -42,10 +42,14 @@ define('DB_HOSTNAME',  'localhost');
 define('DB_USERNAME', 'root');
 define('DB_PASSWORD', 'root');
 define('DB_DATABASE', 'qv');
+define('DB_PORT', '3306');
 
 /*
  * External
  */
+
+// Enabled mailing model
+define('MAIL_ENABLED', true);
 
 // Contact mail
 define('CONTACT_EMAIL', 'info@mydomain.com');